+++ lambda/src/main/scala/pricemigrationengine/model/ZuoraRatePlan.scala
@@ -1,0 +1,20 @@
+package pricemigrationengine.model
+
+import pricemigrationengine.model.OptionReader // not sure why this import is needed as should be visible implicitly
+import upickle.default._
+
+case class ZuoraRatePlan(
+    id: String,
+    productName: String,
+    productRatePlanId: String,
+    ratePlanName: String,
+    ratePlanCharges: List[ZuoraRatePlanCharge],
+    lastChangeType: Option[String] = None
+)
+
+object ZuoraRatePlan {
+  implicit val rw: ReadWriter[ZuoraRatePlan] = macroRW
+
+  def ratePlan(subscription: ZuoraSubscription, ratePlanCharge: ZuoraRatePlanCharge): Option[ZuoraRatePlan] =
+    subscription.ratePlans.find(_.ratePlanCharges.exists(_.number == ratePlanCharge.number))
+}
