+++ lambda/src/main/scala/pricemigrationengine/model/ZuoraSubscription.scala
@@ -21,72 +21,3 @@ case class ZuoraSubscription(
 object ZuoraSubscription {
   implicit val rwSubscription: ReadWriter[ZuoraSubscription] = macroRW
 }
-
-case class ZuoraRatePlan(
-    id: String,
-    productName: String,
-    productRatePlanId: String,
-    ratePlanName: String,
-    ratePlanCharges: List[ZuoraRatePlanCharge],
-    lastChangeType: Option[String] = None
-)
-
-object ZuoraRatePlan {
-  implicit val rw: ReadWriter[ZuoraRatePlan] = macroRW
-
-  def ratePlan(subscription: ZuoraSubscription, ratePlanCharge: ZuoraRatePlanCharge): Option[ZuoraRatePlan] =
-    subscription.ratePlans.find(_.ratePlanCharges.exists(_.number == ratePlanCharge.number))
-}
-
-case class ZuoraRatePlanCharge(
-    productRatePlanChargeId: ZuoraProductRatePlanChargeId,
-    name: String,
-    number: String,
-    currency: Currency,
-    price: Option[BigDecimal] = None,
-    billingPeriod: Option[String] = None,
-    chargedThroughDate: Option[LocalDate] = None,
-    processedThroughDate: Option[LocalDate] = None,
-    specificBillingPeriod: Option[Int] = None,
-    endDateCondition: Option[String] = None,
-    upToPeriodsType: Option[String] = None,
-    upToPeriods: Option[Int] = None,
-    billingDay: Option[String] = None,
-    triggerEvent: Option[String] = None,
-    triggerDate: Option[LocalDate] = None,
-    discountPercentage: Option[Double] = None,
-    originalOrderDate: Option[LocalDate] = None
-)
-
-object ZuoraRatePlanCharge {
-  implicit val rw: ReadWriter[ZuoraRatePlanCharge] = macroRW
-
-  /** Rate plan charge that corresponds with the given invoice item.
-    */
-  def matchingRatePlanCharge(
-      subscription: ZuoraSubscription,
-      invoiceItem: ZuoraInvoiceItem
-  ): Either[AmendmentDataFailure, ZuoraRatePlanCharge] =
-    (for {
-      ratePlan <- subscription.ratePlans
-      ratePlanCharge <- ratePlan.ratePlanCharges
-      if ratePlanCharge.number == invoiceItem.chargeNumber
-    } yield ratePlanCharge).headOption
-      .toRight(AmendmentDataFailure(s"No matching rate plan charge for invoice item '${invoiceItem.chargeNumber}'"))
-}
-
-case class ZuoraAccount(
-    soldToContact: SoldToContact
-)
-
-object ZuoraAccount {
-  implicit val rwSubscription: ReadWriter[ZuoraAccount] = macroRW
-}
-
-case class SoldToContact(
-    country: String
-)
-
-object SoldToContact {
-  implicit val rwSubscription: ReadWriter[SoldToContact] = macroRW
-}
